Rhino animation events
Added rhino animation events
Moved rhino into a rhino folder
Added the slam vfx to rhino's slam
Opacity on rhino slam sprite might need to be adjusted
