+++ Assets/Rendering/3D Assets/Creatures/Mammals/Rhino/ChargeEdit.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 793411139c5fada4f85d5cbd4492b161
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
