+++ Assets/Rendering/3D Assets/Creatures/Mammals/Rhino/SlamEdit.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9853ad7b11bae884e903dfb8ed626856
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
