--- Assets/Scripts/Creature/Major Creatures/Rhino.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f106fffbcc5742643946f032d9bfb132
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
