+++ Assets/Scripts/Creature/Major Creatures/Rhino.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 86c2cbb6c2a45a64782d87a6fc2bb978
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
