+++ Assets/Scripts/Creature/Major Creatures/Rhino/RhinoAnimationEvents.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 62e859a0e7854df409a03012fed7e1a1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
